Factor out common html template code
